started isolating native physics
